networth added and UserStatsModel

+++ CookieData/Entities/Upgrade.cs
@@ -5,10 +5,10 @@ namespace CookieData.Entities
     public class Upgrade : BaseEntity
     {
         public string Name { get; set; }
-        public long Price { get; set; }
-        public int BaseTick { get; set; }
-        public int Level { get; set; }
-        public int Amount { get; set; }
+        public ulong Price { get; set; }
+        public ulong BaseTick { get; set; }
+        public ulong Level { get; set; }
+        public ulong Amount { get; set; }
         public string Image { get; set; }
         public int GameAccountId { get; set; }
         [JsonIgnore]
