+++ CookieData/Model/UpgradeModel.cs
@@ -4,10 +4,10 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public long Price { get; set; }
-        public int BaseTick { get; set; }
-        public int Level { get; set; }
-        public int Amount { get; set; }
+        public ulong Price { get; set; }
+        public ulong BaseTick { get; set; }
+        public ulong Level { get; set; }
+        public ulong Amount { get; set; }
         public string Image { get; set; }
         public int GameAccountId { get; set; }
     }
